+++ middleware/error-handler.js
@@ -4,7 +4,7 @@ module.exports = {
     if (err instanceof Error) {
       res.status(err.status || 500).json({
         status: 'error',
-        message: `${err.name}: ${err.message}`
+        message: err.message
       })
     } else {
       res.status(500).json({
